generic config implementation
